Refactor service protocols
